# contracts/openapi/v1/testforge.yaml
openapi: 3.1.0
...
info:
  title: TestForge Control Plane API
  version: 1.0.0
  description: |
    TestForge 项目、用例、Workflow、Run、Worker、Attempt、报告与诊断的公开契约。
    MySQL 是运行状态真相源；派发和回调均按至少一次语义处理。
    AI 诊断仅给出带证据的建议，不改变确定性测试结果。
servers:
  - url: http://localhost:8081
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/runs:
    post:
      tags: [Runs]
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Run 已幂等创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunSummary"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
  /api/v1/runs/{runId}:
    get:
      tags: [Runs]
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run 聚合及计数
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runs/{runId}/cancel:
    post:
      tags: [Runs]
      operationId: cancelRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRunRequest"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        "409":
          $ref: "#/components/responses/StateConflict"
...
  /api/v1/reports/runs/{runId}:
    get:
      tags: [Reports]
      operationId: getRunReport
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: 确定性运行报告
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunReport"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
...
    Accepted:
      description: 操作已接受
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestResponse"
    ValidationError:
      description: 字段或状态输入非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    IdempotencyConflict:
      description: 相同幂等键对应不同内容
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    StateConflict:
      description: 状态或版本已变化
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Uuid:
      type: string
      format: uuid
    RestResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
          pattern: "^[a-f0-9]{32}$"
        data: {}
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/RestResponse"
        - type: object
          required: [code, message, traceId]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - RESOURCE_NOT_FOUND
                - IDEMPOTENCY_CONFLICT
                - STATE_CONFLICT
                - LEASE_EXPIRED
                - CONCURRENCY_LIMITED
                - DISPATCH_UNAVAILABLE
                - AI_PROVIDER_UNAVAILABLE
                - TARGET_SOURCE_NOT_CONFIGURED
                - REVISION_NOT_FOUND
                - REVISION_RESOLUTION_FAILED
                - REVISION_RESOLUTION_TIMEOUT
            details:
              type: object
              additionalProperties: true
...
    TaskRevisionOverride:
      type: object
      additionalProperties: false
      required: [workflowNodeId, revision]
      properties:
        workflowNodeId:
          $ref: "#/components/schemas/Uuid"
        revision:
          $ref: "#/components/schemas/RevisionSelector"
...
    RevisionSelector:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          enum: [DEFAULT_BRANCH, BRANCH, TAG, COMMIT]
        value:
          type: string
          maxLength: 255
          nullable: true
...
    TaskTargetRevision:
      type: object
      additionalProperties: false
      required: [repositoryUrl, requestedType, resolvedCommit, resolvedAt]
      properties:
        repositoryUrl:
          type: string
          maxLength: 2048
        requestedType:
          type: string
          enum: [DEFAULT_BRANCH, BRANCH, TAG, COMMIT]
        requestedValue:
          type: [string, "null"]
        resolvedCommit:
          type: string
          pattern: "^[0-9a-f]{40}$"
        resolvedAt:
          type: string
          format: date-time
...
    CreateRunRequest:
      type: object
      additionalProperties: false
      required: [projectId, targetId, environmentId, workflowId, workflowVersion, priority, maxConcurrency, requestKey]
      properties:
        projectId:
          $ref: "#/components/schemas/Uuid"
        targetId:
          $ref: "#/components/schemas/Uuid"
        environmentId:
          $ref: "#/components/schemas/Uuid"
        workflowId:
          $ref: "#/components/schemas/Uuid"
        workflowVersion:
          type: integer
          minimum: 1
        priority:
          type: integer
          minimum: 0
          maximum: 9
        maxConcurrency:
          type: integer
          minimum: 1
          maximum: 20
        processConcurrency:
          type: integer
          minimum: 1
          maximum: 100
          description: PROCESS_POOL 独立并发上限；省略时继承 maxConcurrency。
        deviceConcurrency:
          type: integer
          minimum: 1
          maximum: 100
          description: EXCLUSIVE_DEVICE 独立并发上限；省略时继承 maxConcurrency。
        requestKey:
          $ref: "#/components/schemas/Uuid"
        targetRevision:
          $ref: "#/components/schemas/RevisionSelector"
        taskRevisionOverrides:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/TaskRevisionOverride"
        deploymentProfileId:
          $ref: "#/components/schemas/Uuid"
          description: 指定后，Run 的 Task 在该 Commit 部署 READY 前保持 WAITING_DEPLOYMENT。
...
    CancelRunRequest:
      type: object
      additionalProperties: false
      required: [requestKey, reason]
      properties:
        requestKey:
          $ref: "#/components/schemas/Uuid"
        reason:
          type: string
          minLength: 1
          maxLength: 500
    RunState:
      type: string
      enum: [QUEUED, RUNNING, CANCELLING, SUCCEEDED, FAILED, COMPLETED_WITH_WARNINGS, CANCELLED]
...
    RunSummary:
      type: object
      required: [id, state, totalTasks, taskCounts, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        comparisonGroupId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        testJobId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        jobConfigVersion:
          type: integer
          minimum: 0
        state:
          $ref: "#/components/schemas/RunState"
        totalTasks:
          type: integer
          minimum: 0
        taskCounts:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
...
    RunReport:
      type: object
      required: [runId, runState, total, passed, failed, pending, durationMs, p50DurationMs, p95DurationMs, failureCategories, results, artifactKeys]
      properties:
        runId:
          $ref: "#/components/schemas/Uuid"
        runState:
          $ref: "#/components/schemas/RunState"
        total: { type: integer, minimum: 0 }
        passed: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
        pending: { type: integer, minimum: 0 }
        durationMs: { type: integer, format: int64, minimum: 0 }
        p50DurationMs:
          type: integer
          minimum: 0
        p95DurationMs:
          type: integer
          minimum: 0
        failureCategories:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
        results:
          type: array
          items:
            $ref: "#/components/schemas/RunReportResult"
        artifactKeys:
          type: array
          items: { type: string }
    RunReportResult:
      type: object
      required: [taskId, caseId, attemptId, status, durationMs, summary, artifactKeys, targetRevision]
      properties:
        taskId: { $ref: "#/components/schemas/Uuid" }
        caseId: { $ref: "#/components/schemas/Uuid" }
        attemptId: { $ref: "#/components/schemas/Uuid" }
        status: { type: string, enum: [PASSED, ASSERTION_FAILED, INFRA_FAILED, CANCELLED] }
        durationMs: { type: integer, format: int64, minimum: 0 }
        failureType: { type: [string, "null"] }
        summary: { type: string }
        artifactKeys:
          type: array
          items: { type: string }
        targetRevision:
          oneOf:
            - $ref: "#/components/schemas/TaskTargetRevision"
            - type: "null"
...